split component into different files
https://github.com/multiversx/mx-explorer-dapp/pull/104#discussion_r1628873133
